Use only simple fields from STK
* Force STK io broker to use simple fields.
  Otherwise, vector fields are declared with Cartesian tag.
* Use simple fields in all meta data structs of STK meshes
* Fixed STK field containers handling of STK fields
  Now that we use simple fields, there is no field_array_rank in stk
  fields anymore, and there is no distinction between "scalar" and
  "vector" fields (in terms of static type)
* In Ioss mesh, cannot query field array rank of a stk Field anymore
* In STKDiscretization, metaData's get_field only needs the data type as
  template arg
