+++ src/disc/stk/Albany_STKDiscretization.cpp
@@ -347,8 +347,7 @@ STKDiscretization::transformMesh()
     stkMeshStruct->PBCStruct.scale[0] *= L;
     stkMeshStruct->PBCStruct.scale[1] *= L;
     stk::mesh::Field<double>* surfaceHeight_field =
-        metaData->get_field<stk::mesh::Field<double>>(
-            stk::topology::NODE_RANK, "surface_height");
+        metaData->get_field<double>(stk::topology::NODE_RANK, "surface_height");
     const auto numOverlapNodes = getLocalSubdim(getOverlapNodeVectorSpace());
     for (int i = 0; i < numOverlapNodes; i++) {
       double* x = stk::mesh::field_data(*coordinates_field, overlapnodes[i]);
@@ -378,8 +377,7 @@ STKDiscretization::transformMesh()
     stkMeshStruct->PBCStruct.scale[0] *= L;
     stkMeshStruct->PBCStruct.scale[1] *= L;
     stk::mesh::Field<double>* surfaceHeight_field =
-        metaData->get_field<stk::mesh::Field<double>>(
-            stk::topology::NODE_RANK, "surface_height");
+        metaData->get_field<double>(stk::topology::NODE_RANK, "surface_height");
     const auto numOverlapNodes = getLocalSubdim(getOverlapNodeVectorSpace());
     for (int i = 0; i < numOverlapNodes; i++) {
       double* x = stk::mesh::field_data(*coordinates_field, overlapnodes[i]);
@@ -410,8 +408,7 @@ STKDiscretization::transformMesh()
     stkMeshStruct->PBCStruct.scale[0] *= L;
     stkMeshStruct->PBCStruct.scale[1] *= L;
     stk::mesh::Field<double>* surfaceHeight_field =
-        metaData->get_field<stk::mesh::Field<double>>(
-            stk::topology::NODE_RANK, "surface_height");
+        metaData->get_field<double>(stk::topology::NODE_RANK, "surface_height");
     const auto numOverlapNodes = getLocalSubdim(getOverlapNodeVectorSpace());
     for (int i = 0; i < numOverlapNodes; i++) {
       double* x = stk::mesh::field_data(*coordinates_field, overlapnodes[i]);
@@ -430,8 +427,7 @@ STKDiscretization::transformMesh()
     stkMeshStruct->PBCStruct.scale[0] *= L;
     stkMeshStruct->PBCStruct.scale[1] *= L;
     stk::mesh::Field<double>* surfaceHeight_field =
-        metaData->get_field<stk::mesh::Field<double>>(
-            stk::topology::NODE_RANK, "surface_height");
+        metaData->get_field<double>(stk::topology::NODE_RANK, "surface_height");
     const auto numOverlapNodes = getLocalSubdim(getOverlapNodeVectorSpace());
     for (int i = 0; i < numOverlapNodes; i++) {
       double* x = stk::mesh::field_data(*coordinates_field, overlapnodes[i]);
@@ -450,8 +446,7 @@ STKDiscretization::transformMesh()
     stkMeshStruct->PBCStruct.scale[0] *= L;
     stkMeshStruct->PBCStruct.scale[1] *= L;
     stk::mesh::Field<double>* surfaceHeight_field =
-        metaData->get_field<stk::mesh::Field<double>>(
-            stk::topology::NODE_RANK, "surface_height");
+        metaData->get_field<double>(stk::topology::NODE_RANK, "surface_height");
     const auto numOverlapNodes = getLocalSubdim(getOverlapNodeVectorSpace());
     for (int i = 0; i < numOverlapNodes; i++) {
       double* x = stk::mesh::field_data(*coordinates_field, overlapnodes[i]);
@@ -475,8 +470,7 @@ STKDiscretization::transformMesh()
     stkMeshStruct->PBCStruct.scale[0] *= L;
     stkMeshStruct->PBCStruct.scale[1] *= L;
     stk::mesh::Field<double>* surfaceHeight_field =
-        metaData->get_field<stk::mesh::Field<double>>(
-            stk::topology::NODE_RANK, "surface_height");
+        metaData->get_field<double>(stk::topology::NODE_RANK, "surface_height");
     const auto numOverlapNodes = getLocalSubdim(getOverlapNodeVectorSpace());
     for (int i = 0; i < numOverlapNodes; i++) {
       double* x = stk::mesh::field_data(*coordinates_field, overlapnodes[i]);
@@ -506,8 +500,7 @@ STKDiscretization::transformMesh()
 #endif
     stkMeshStruct->PBCStruct.scale[0] *= L;
     stk::mesh::Field<double>* surfaceHeight_field =
-        metaData->get_field<stk::mesh::Field<double>>(
-            stk::topology::NODE_RANK, "surface_height");
+        metaData->get_field<double>(stk::topology::NODE_RANK, "surface_height");
     const auto numOverlapNodes = getLocalSubdim(getOverlapNodeVectorSpace());
     for (int i = 0; i < numOverlapNodes; i++) {
       double* x = stk::mesh::field_data(*coordinates_field, overlapnodes[i]);
